maritime_backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
